Try to fix NPY file parsing by upgrading to Nom 5
I am not sure how this broke because I haven't monkied with this part of the
code for a long time. But, parsing `meanbp-c.npy` was exiting with an
`Incomplete(Size(1))` error. (Maybe I've only been using `meanbp-c3bit.npy`
and that doesn't trigger the problem?)

This probably could have been solved with staying at Nom 4, but it's good to
stay up-to-date anyway, so I bumped to Nom 5, which does seem to have fixed
the problem. It looks like we're supposed to make much more substantial
changes for Nom 5, from macros to functions, but this looks to get the job
done. I do get a deprecation warning:

```
warning: use of deprecated item 'ws': whitespace parsing only works with macros and will not be updated anymore
```

To be dealt with later. Note that `npy-rs`, which which I stole the parser,
still uses Nom 3.